+++ docs/history.rst
@@ -2,7 +2,9 @@
 
 Version History
 ===============
-*next release*
+*next*
+  * Environment values FORCE_COLOR_, CLICOLOR_FORCE_, and NO_COLOR_ overrides
+    :class:`blessed.Terminal` argument value, ``force_styling``.
   * performance improvements in :meth:`~Terminal.rgb_downconvert` for 256-color
     terminals, :ghpull:`298`.
 
@@ -339,3 +341,7 @@ Version History
 .. _`API Documentation`: http://blessed.rtfd.org
 .. _`PDCurses`: https://www.lfd.uci.edu/~gohlke/pythonlibs/#curses
 .. _`ansi`: https://github.com/tehmaze/ansi
+.. _FORCE_COLOR: https://force-color.org/
+.. _CLICOLOR_FORCE: https://bixense.com/clicolors/
+.. _NO_COLOR: https://no-color.org/
+